refactor(agent-service): group DataGuard tests under __tests__/
The dataguard/ folder had 8 source + 8 test files in one directory and was
getting hard to scan. Move the seven DataGuard-specific test files into a
__tests__/ subdirectory and update their relative imports (one extra ../).
The types/dataguard.test.ts stays put — it's in src/types/, not under
dataguard/, and that folder isn't crowded. bun test still auto-discovers;
all 159 tests still pass.
